server/requirements.txt
flask
pymongo
python-dotenv
openai